role permissions improv.

+++ app/Models/Role.php
@@ -11,8 +11,27 @@ class Role extends SpatieRoleModel
 
     public function delete()
     {
+        if($this->name === 'admin' || $this->name === 'super user') return;
+
         $this->permissions()->detach();
+        $this->users()->detach();
+
         parent::delete();
+    }
+
+
+    public function isAdmin()
+    {
+        return $this->name === 'admin' || $this->name === 'super user';
+    }
+
+
+    /**
+     * Exclude the 'super user' role
+     */
+    public static function allExceptSU()
+    {
+        return self::where('name', '!=', 'super user')->get();
     }
 
 }
